pyproject.toml
[project]
name = "trajectoire-pyro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.6.2",
    "autogen-ext[open-ai,openai]>=0.6.2",
    "ipykernel>=6.29.5",
    "pandas>=2.3.0",
    "pygithub>=2.6.1",
    "pyro-ppl>=1.9.1",
    "pytest>=8.4.1",
    "torch>=2.7.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]                 # tell setuptools that sources live in src/
package-dir = {"" = "src"}

[tool.setuptools.packages.find]   # auto‑discover packages under src/
where = ["src"]

[project.scripts]
owner = "trajpyro.agents.owner:main"
smoke = "trajpyro.smoke:main"